Trigger APF program update on IPv6 address state change
Modified handleLinkPropertiesUpdate() to call
mApfFilter.setLinkProperties() whenever link addresses change, ensuring
the APF program is updated when an IPv6 address transitions from
tentative to non-tentative state. This may result in redundant calls to
mApfFilter.setLinkProperties() but program will not be regenerate in the
second call as there is no if no actual address change occurs.

When calling the setLinkProperties(), IpClient always called
assembleLinkProperties() first and then pass in the new newLp. The
mLinkProperties variable is never passed into APF.

Test: TH, manual test
Change-Id: Id4bdf6d26654221bb2d5aa83cbd039520f4fae9f