Register IpClient as OffloadEngine when APF supports mDNS offload
Updated IpClient to register itself as an OffloadEngine instance when
the APF indicates support for mDNS offload functionality.

The offloadEngine is registered when IpClient starts running and
unregistered when it stops. However, the offloadEngine is also
registered if the IpClient's ApfCapabilities are updated. This can
happen when a WiFi STA starts in secondary mode and then switches
to primary mode. If WiFi STA starts in secondary mode, then
ApfCapabilities is null so there is no ApfFilter created when IpClient
entered the RunningState.

Test: TH
Change-Id: I4a0c73a1ebdcd306f981236c0298b48414a563c4