Support non-IP address lookups in FakeDns.
This is a no-op refector which will allow future tests to use
DNS lookup types other than A and AAAA (e.g., SVCB). This
requires the ability to call the DnsResolver.Callback type that
takes a byte[] instead of a List<InetAddress>.

The CL makes the FakeDns / AnswerSupplier test classes use
List<String> instead of List<InetAddress>, and converts the
string to the type required by the callback (currently
List<InetAddress>, but in the future byte[] as well, depending
on the query type) just prior to calling the callback.

Bug: 240259333
Test: test-only change
Change-Id: I32b35c43e5a0ec262289120a6da371613aea0cfd