Default enable ApfFilter for all devices < V
The ApfFilter experiment has previously been at 50% public on Pixel.
There was a slight regression on provisioning latency ~30ms for both
IPv4 and IPv6, though the regression did not seem to stem from a bug but
rather some inefficiency in the code.

The ApfFilter experiment was ramped back down to 0% due to a bug that
dropped broadcasted DHCP replies with the DF flag set. This bug was
fixed in both V FRC prebuilts and the August mainline train (which has
since rolled out to 100%). Given that the experiment has rolled out to
2% public again, and the November module roll out schedule still allows
for this to be reverted.

However, given that this code has been written over a year ago, and we
keep (ab)using the existing experiment flag to launch various other
"improvements" which should really be behind their own flag, this really
needs to get launched ASAP.

Note that programs are still capped at 1024 before V. While the
LegacyApfFilter almost never generated programs of this size, it
seems unlikely that this will break (given that the lowest program
size we know to be broken is 1950).

Test: n/a
Change-Id: Ia55324a2af0afd7b399a71458273d2afd7920aa7